08 Java/JUnit_app/src/main/java/org/models/Cuenta.java
package org.models;


import java.math.BigDecimal;

public class Cuenta {

    private String persona;
    private BigDecimal saldo;

    public Cuenta(String persona, BigDecimal saldo) {
        this.persona = persona;
        this.saldo = saldo;
    }

    public String getPersona() {
        return persona;
    }

    public void setPersona(String persona) {
        this.persona = persona;
    }

    public BigDecimal getSaldo() {
        return saldo;
    }

    public void setSaldo(BigDecimal saldo) {
        this.saldo = saldo;
    }

    public void credito(BigDecimal saldo){
    this.saldo = this.saldo.add(saldo);
    }
    public void debito(BigDecimal saldo){

        BigDecimal nuevoSaldo = this.saldo.subtract(saldo);

        if(nuevoSaldo.compareTo(BigDecimal.ZERO) < 0){
            throw new DineroInsuficienteException("Dinero insuficiente");
        }
        this.saldo = nuevoSaldo;

    }

}
